shuffling distance stuff
